+++ update-version.sh
@@ -3,8 +3,8 @@
 set -e
 
 MAJOR=1
-MINOR=9
-PATCH=5
+MINOR=11
+PATCH=0
 FULL=${MAJOR}.${MINOR}.${PATCH}
 
 # For compatibility between OSX and Unix, we need to set a backup file and then delete it